Implementação Habilidades Espada Parte 2
-implementação da habilidade corte rápido
-implementação da habilidade investida real
-atualização sistema de save + habilidades

+++ Assets/Scripts/Habilidades/HabilidadesClasse/Guerreiro/GolpeEspartano/HabilidadeGolpeEspartanoNv3.cs
@@ -12,7 +12,7 @@ public class HabilidadeGolpeEspartanoNv3 : HabilidadeAtiva
     [SerializeField]
     private float raioDeDistanciaDoTaunt;
     [SerializeField]
-    private int tempoDeStun = 3;
+    private float tempoDeStun = 3;
     public GameObject vfx;
 
     public override void AtivarEfeito(IAPersonagemBase personagem)
